requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=0.10.0
torchtext>=0.11.0
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
pillow>=8.2.0
opencv-python>=4.5.0
albumentations>=1.3.1

# Development dependencies
pytest>=7.0.0
